refactor render out of trackgen
