+++ main.go
@@ -1,7 +1,7 @@
 package main
 
-import "github.com/socialgorithm/elon-server/track"
+import "github.com/socialgorithm/elon-server/render"
 
 func main() {
-	track.Main()
+	render.Main()
 }
