+++ render/constants.go
@@ -1,0 +1,14 @@
+package render
+
+import "golang.org/x/image/colornames"
+
+const width = 1024
+const height = 768
+const lineThickness = 4
+const segmentLength = 15
+
+var bgColor = colornames.Green
+var roadColor = colornames.Gray
+var outerLinesColor = colornames.White
+var centerLineColor = colornames.Yellow
+var segmentColor = colornames.Red
